interfaces/tls_certificates/pyproject.toml
[project]
name = "charmlibs-interfaces-tls_certificates"
description = "The charmlibs.interfaces.tls_certificates package."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name="The TLS team at Canonical"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 5 - Production/Stable",
]
dynamic = ["version"]
dependencies = [
    "cryptography>=43.0.0",
    "jsonschema",
    "ops",
    "pydantic",
]

[dependency-groups]
lint = [  # installed for `just lint interfaces/tls_certificates` (unit, functional, and integration are also installed)
    "typing_extensions",
]
unit = [  # installed for `just unit interfaces/tls_certificates`
    "ops[testing]",
    "pkilint",
]
functional = [  # installed for `just functional interfaces/tls_certificates`
]
integration = [  # installed for `just integration interfaces/tls_certificates`
    "jubilant",
    "juju>3",
    "pytest-operator",
    "pytest-asyncio==0.21.2",
]

[project.urls]
"Repository" = "https://github.com/canonical/charmlibs"
"Issues" = "https://github.com/canonical/charmlibs/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/charmlibs"]

[tool.hatch.version]
path = "src/charmlibs/interfaces/tls_certificates/_version.py"

[tool.ruff]
extend = "../../pyproject.toml"
src = ["src", "tests/unit", "tests/functional", "tests/integration"]  # correctly sort local imports in tests

[tool.ruff.lint.extend-per-file-ignores]
"src/charmlibs/interfaces/tls_certificates/_tls_certificates.py" = [
    "B904",  # FIXME: use raise ... from
    "B909",  # FIXME: mutation to loop iterable during iteration
    "PERF203",  # FIXME: ignore inline or refactor
]
"tests/*" = [
    "E501",  # line too long
]

[tool.pyright]
extends = "../../pyproject.toml"
include = ["src", "tests"]
pythonVersion = "3.10"  # check no python > 3.10 features are used

[tool.pytest.ini_options]
markers = [
    "upgrade: test that refreshes the chamr with the new library version",
]

[tool.charmlibs.functional]
ubuntu = []  # ubuntu versions to run functional tests with, e.g. "24.04" (defaults to just "latest")
pebble = []  # pebble versions to run functional tests with, e.g. "v1.0.0", "master" (defaults to no pebble versions)
sudo = false  # whether to run functional tests with sudo (defaults to false)

[tool.charmlibs.integration]
# tags to run integration tests with (defaults to running once with no tag, i.e. tags = [''])
# Available in CI in tests/integration/pack.sh and integration tests as CHARMLIBS_TAG
tags = []  # Not used by the pack.sh and integration tests generated by the template
